Update(plotting): Bug fixes and simplifications
* Refactored plot_bc code to remove the stress_period_data.to_array() method from plot_bc(). This now sets this data directly from the mflist record.
* Restored CHD plotting color to blue
* Simplifications to the PlotUtilities call stack.
* Merged _StructuredMapView, _VertexMapView, and _UnstructuredMapView into the PlotMapView class to reduce the amount of indirection and the size of the call stack. 
* Removed umap.py, vmap.py
* Fixed crosssection.py mf6 looking for nonexistent laycbd variable
