+++ flopy/plot/__init__.py
@@ -21,6 +21,6 @@
 
     """
 from .plotutil import SwiConcentration, plot_shapefile, shapefile_extents, PlotUtilities
-from .map import ModelMap
+from .map import ModelMap, PlotMapView
 from .crosssection import ModelCrossSection
-from .plotbase import PlotMapView, PlotCrossSection
+from .plotbase import PlotCrossSection
